kernel: Add debug system
The debug system is initialized first to enable early-boot processes to
log messages.  On qemu-virt-riscv, the NS16550A device provides both the
debug and console interfaces.  The NS16550A implementation is split into
debug and device sections, sharing a common low-level interface to the
device itself.  The debug implementation is intentionally sparse, making
a best-effort attempt to write messages as they are given.  The device
implementation provides the console.

Panic now logs debug messages instead of console messages in case of
early-boot panic.  A new printf function, dprintf, outputs to the debug
log.